Create a Probot middleware
